refactor(desktop): split-pane settings/projects with unified detail route
Collapses the separate /settings/project and /settings/v2-project
trees into a single /settings/projects layout. Left pane lists v1 + v2
projects (deduped as before); right pane renders the selected
project's detail at /settings/projects/\$projectId. The detail route
branches on v2 cloud-collection membership and renders the existing
ProjectSettings or V2ProjectSettings component.

Also removes the in-page "Back to Projects" button from the project
settings header — the sidebar already provides navigation. Repoints
every caller (workspace sidebar, dashboard sidebar, workspace run
button, v2 conflict redirect, settings-sidebar matchRoute) to the new
URL. Old route files deleted; cloud/secrets tree left in place as
orphan (no external callers).
