+++ apps/desktop/src/renderer/routes/_authenticated/_dashboard/components/TopBar/components/WorkspaceRunButton/WorkspaceRunButton.tsx
@@ -57,7 +57,7 @@ export const WorkspaceRunButton = memo(function WorkspaceRunButton({
 		if (!hasRunCommand && projectId) {
 			setSettingsSearchQuery("scripts");
 			void navigate({
-				to: "/settings/project/$projectId/general",
+				to: "/settings/projects/$projectId",
 				params: { projectId },
 			});
 			return;
