+++ apps/desktop/src/renderer/routes/_authenticated/settings/project/$projectId/components/ProjectSettingsHeader/ProjectSettingsHeader.tsx
@@ -1,7 +1,4 @@
-import { Button } from "@superset/ui/button";
-import { Link } from "@tanstack/react-router";
 import type { ReactNode } from "react";
-import { HiArrowLeft } from "react-icons/hi2";
 
 interface ProjectSettingsHeaderProps {
 	title: string;
@@ -13,18 +10,9 @@ export function ProjectSettingsHeader({
 	children,
 }: ProjectSettingsHeaderProps) {
 	return (
-		<div className="mb-8 space-y-4">
-			<Button variant="ghost" size="sm" asChild>
-				<Link to="/settings/projects">
-					<HiArrowLeft className="h-4 w-4" />
-					Projects
-				</Link>
-			</Button>
-
-			<div>
-				<h2 className="text-xl font-semibold">{title}</h2>
-				{children && <div className="mt-1">{children}</div>}
-			</div>
+		<div className="mb-8">
+			<h2 className="text-xl font-semibold">{title}</h2>
+			{children && <div className="mt-1">{children}</div>}
 		</div>
 	);
 }
