+++ apps/desktop/src/renderer/routes/_authenticated/settings/projects/components/ProjectsSettingsSidebar/index.ts
@@ -1,0 +1,1 @@
+export { ProjectsSettingsSidebar } from "./ProjectsSettingsSidebar";
