+++ apps/desktop/src/renderer/routes/_authenticated/settings/v2-project/$projectId/components/V2ProjectSettings/components/ProjectLocationSection/ProjectLocationSection.tsx
@@ -186,7 +186,7 @@ export function ProjectLocationSection({
 								const target = conflict;
 								setConflict(null);
 								navigate({
-									to: "/settings/v2-project/$projectId/general",
+									to: "/settings/projects/$projectId",
 									params: { projectId: target.id },
 								});
 							}}
